Move initialize function for geopotential
